refactor: `HandleDragManager` -> `HandleResizeManager`
